refactor(mitm-addon): split auth base forwarder
